# rest_client/openapi.yaml
openapi: 3.0.3
info:
  title: rogaining-system
  description: rogaining-systemのAPI設計
  version: 0.1.1
servers:
- url: http://localhost
  description: local
- url: http://127.0.0.1:3000
  description: local sam
- url: https://rjdykda599.execute-api.ap-northeast-1.amazonaws.com/dev/
  description: dev
paths:
  /admin/create_user/:
    post:
      tags:
      - admin
      summary: crete login user
      operationId: creteLoginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateAdmin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/login/:
    get:
      tags:
      - admin
      summary: get login disp
      operationId: getLoginDisp
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
    post:
      tags:
      - admin
      summary: post login
      operationId: postLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginAdmin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/scores/:
    get:
      tags:
      - admin
      summary: get scores
      operationId: getScores
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
      security:
      - OAuth2PasswordBearerOrCookie: []
  /card/card_token/:
    post:
      tags:
      - card
      - admin
      summary: generate card token
      operationId: generateCardToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputCardToken'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearerOrCookie: []
  /card/:
    post:
      tags:
      - card
      - seal
      summary: attach card
      description: Creates a new Post
      operationId: attachCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachCardForm'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /lineapi/callback:
    post:
      summary: line messaging
      description: ユーザーからメッセージが送信された際、LINE Message APIからこちらのメソッドが呼び出されます。
      operationId: lineMessaging
      parameters:
      - name: x-line-signature
        in: header
        required: false
        schema:
          title: X-Line-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /places/:
    get:
      tags:
      - admin
      summary: get places
      operationId: getPlaces
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
      security:
      - OAuth2PasswordBearerOrCookie: []
    post:
      tags:
      - admin
      summary: post places
      description: get all place
      operationId: postPlaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlacesInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PlaceOutput'
                type: array
                title: Response Post Places Places  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /seal/:
    get:
      tags:
      - seal
      summary: seal stamp disp
      description: 'return seal dispray:


        - **username**(str):

        - **password**(str):

        - **sex**(int): [男性:0,女性:1,その他2]

        - **age**(int):'
      operationId: sealStampDisp
      parameters:
      - name: place_id
        in: query
        required: false
        schema:
          type: string
          title: Place Id
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - seal
      summary: seal stamp post
      description: Creates a new Post
      operationId: sealStampPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StampSealForm'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AdminOut:
      properties:
        username:
          type: string
          title: Username
          description: user名
        id:
          type: integer
          title: Id
          description: id
      type: object
      required:
      - username
      - id
      title: AdminOut
    AttachCardForm:
      properties:
        lineToken:
          type: string
          title: Linetoken
          description: liff用のToken
        card_token:
          type: string
          title: Card Token
          description: カードのトークン
      type: object
      required:
      - lineToken
      - card_token
      title: AttachCardForm
    CardTokenResponse:
      properties:
        cardToken:
          type: string
          title: Cardtoken
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      required:
      - cardToken
      - name
      title: CardTokenResponse
    GenerateAdmin:
      properties:
        username:
          type: string
          title: Username
          description: user名
        password:
          type: string
          title: Password
          description: パスワード
        token:
          type: string
          title: Token
          description: トークン
      type: object
      required:
      - username
      - password
      - token
      title: GenerateAdmin
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InputCardToken:
      properties:
        team_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Team Name
          description: スタンプを押した場所
      type: object
      title: InputCardToken
    LoginAdmin:
      properties:
        username:
          type: string
          title: Username
          description: user名
        password:
          type: string
          title: Password
          description: パスワード
      type: object
      required:
      - username
      - password
      title: LoginAdmin
    Place:
      properties:
        name:
          type: string
          title: Name
          description: 名前
        score:
          type: integer
          title: Score
          description: スコア
        gpsLatitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Gpslatitude
          description: 緯度
        gpsLongitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Gpslongitude
          description: 経度
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
          description: 有効か
          default: true
      type: object
      required:
      - name
      - score
      title: Place
    PlaceOutput:
      properties:
        name:
          type: string
          title: Name
          description: 名前
        score:
          type: integer
          title: Score
          description: スコア
        gpsLatitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Gpslatitude
          description: 緯度
        gpsLongitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Gpslongitude
          description: 経度
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
          description: 有効か
          default: true
        id:
          type: string
          format: uuid4
          title: Id
      type: object
      required:
      - name
      - score
      - id
      title: PlaceOutput
    PlacesInput:
      properties:
        places:
          items:
            $ref: '#/components/schemas/Place'
          type: array
          title: Places
      type: object
      required:
      - places
      title: PlacesInput
    StampSealForm:
      properties:
        lineToken:
          type: string
          title: Linetoken
          description: liff用のToken
        place_id:
          type: string
          title: Place Id
          description: スタンプを押した場所
      type: object
      required:
      - lineToken
      - place_id
      title: StampSealForm
      description: スタンプの押印処理時に受け取るpostのform的なもの
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
      - access_token
      - token_type
      title: Token
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearerOrCookie:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token
tags:
- name: admin
  description: 管理系のAPI
- name: user
  description: ユーザーに関するAPI
- name: seal
  description: スタンプを押す用のAPIとか
- name: card
  description: card系のAPI, token生成や